refactor(database): move issueCounts #891
On-behalf-of: SAP Michal Krzyz <michal.krzyz@sap.com>
